+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2020-11-05
+### Added
+- Install AWS Elastic Beanstalk CLI
+- Install Serverless
+- `gtc` command for committing on trunk based development
+
 ## [1.3.0] - 2020-11-02
 ### Added
 - Install Arduino IDE